ToyPublicKeys

Documentation for ToyPublicKeys.

ToyPublicKeys.RSAPrivateKey

ToyPublicKeys.RSAPublicKey

ToyPublicKeys.RSAStep

ToyPublicKeys.RSAStep

ToyPublicKeys.RSAStep

ToyPublicKeys.RSAStep

ToyPublicKeys.decrypt

ToyPublicKeys.decrypt

ToyPublicKeys.encrypt

ToyPublicKeys.generate_rsa_key_pair

ToyPublicKeys.is_padded

ToyPublicKeys.is_probab_prime_p

ToyPublicKeys.pad

ToyPublicKeys.pad

ToyPublicKeys.power_crt

ToyPublicKeys.power_crt

ToyPublicKeys.power_crt_components

ToyPublicKeys.rand_prime_for_rsa

ToyPublicKeys.random_bigint_from_range

ToyPublicKeys.sign

ToyPublicKeys.sign

ToyPublicKeys.unpad

ToyPublicKeys.unpad

ToyPublicKeys.verify_signature

ToyPublicKeys.RSAPrivateKey — Type

RSAPrivateKey is PrivateKey struct for RSA.
It holds all information to derive public key and make efficient calculations.

ToyPublicKeys.RSAPublicKey — Type

RSAPrivateKey is PublicKey struct for RSA.
It holds all neccecary information to perform public key computations, but not more.

ToyPublicKeys.RSAStep — Method

Fast implementation of the RSA exponentiation step, when RSAPrivateKey is provided.
It uses Chinese remainer theorem for very fast `exp() mod n` calculations.

ToyPublicKeys.RSAStep — Method

RSA exponentiation step when only public key is available. 
Uses repeated squares and other fast modulo exponentiation tricks in its GMP implementation.

ToyPublicKeys.RSAStep — Method

RSA exponentiation step for Strings.

ToyPublicKeys.RSAStep — Method

RSA exponentiation step for AbstractVectors (arbitrary buffers).
Same as the core BigInt version but with Vector to BigInt construction
and deconstruction.

ToyPublicKeys.decrypt — Method

RSA decryption function for strings.

ToyPublicKeys.decrypt — Method

RSA decryption function for vectors.

ToyPublicKeys.encrypt — Method

RSA encryption function with random padding.

ToyPublicKeys.generate_rsa_key_pair — Method

RSA key pair constructor.

ToyPublicKeys.is_padded — Method

Checks for magic bytes of PKCS#1 v1.5 padding.

ToyPublicKeys.is_probab_prime_p — Method

Binding for GMPlib's gmpz_probab_prime_p function. 
Only uses probability based tests as it would equall factoring n otherwise.

ToyPublicKeys.pad — Method

Core implementation of the PKCS#1 v1.5 padding.

ToyPublicKeys.pad — Method

Wrapper for the core pad function.

ToyPublicKeys.power_crt — Method

Wrapper around core implementation, only for generating the parameters if they are not provided.

ToyPublicKeys.power_crt — Method

Core implementation of exp mod using CRT.

ToyPublicKeys.power_crt_components — Method

Utility function for calculating the CRT parameters.

ToyPublicKeys.rand_prime_for_rsa — Function

Helper function for rsa random primes, since they should satisfy special properties.

ToyPublicKeys.random_bigint_from_range — Method

Custom random big int function since core julia does not yet provide 'proper one' (in my opinion).

ToyPublicKeys.sign — Method

Sign AbstractVector (arbitrary buffer) with RSA key.

ToyPublicKeys.sign — Method

Sign string with RSA key.

ToyPublicKeys.unpad — Method

Wrapper for the core unpad function.

ToyPublicKeys.unpad — Method

Core implementation for the PKCS#1 v1.5 pad unwrapping.

ToyPublicKeys.verify_signature — Method

Verify the sign result.