ToyPublicKeys.RSAKey — Type

RSAKey

Union of RSAPrivateKey and RSAPublicKey for methods, that do not require specific key.

ToyPublicKeys.RSAPrivateKey — Type

RSAPrivateKey

RSAPrivateKey is PrivateKey struct for RSA. It holds all information to derive public key and make efficient calculations.

ToyPublicKeys.RSAPublicKey — Type

RSAPrivateKey

RSAPrivateKey is PublicKey struct for RSA.

It holds all neccecary information to perform public key computations, but not more.

ToyPublicKeys.RSADP — Method

RSAStep(::pkcs1_v1_5_t, msg::BigInt, key::RSAPublicKey)

RSA exponentiation step when only public key is available. Uses repeated squares and other fast modulo exponentiation tricks in its GMP implementation (Base.GMP.MPZ.powm).

ToyPublicKeys.RSAEP — Method

RSAStep(::pkcs1_v1_5_t, msg::BigInt, key::RSAPrivateKey)

Fast implementation of the RSA exponentiation step when RSAPrivateKey is provided. It uses Chinese remainer theorem for very fast exp() mod n calculations.

ToyPublicKeys.RSASP1 — Method

RSASP1(::pkcs1_v1_5_t, msg::BigInt, key::RSAPrivateKey)

ToyPublicKeys.RSAStep — Method

RSAStep(::pkcs1_v1_5_t, msg::String, key::RSAKey)

RSA exponentiation step for Strings. Only prepares the buffer for RSAStep(msg::BigInt, key::RSAPublicKey).

ToyPublicKeys.RSAStep — Method

RSAStep(::pkcs1_v1_5_t, msg::AbstractVector{T}, key::RSAKey) where {T<:Base.BitInteger}

RSA exponentiation step for AbstractVectors (arbitrary buffers). Only prepares the buffer for RSAStep(msg::BigInt, key::RSAPublicKey).

ToyPublicKeys.RSAVP1 — Method

RSAVP1(::pkcs1_v1_5_t, msg::BigInt, key::RSAPublicKey)

ToyPublicKeys.decrypt — Method

decrypt(::pkcs1_v1_5_t, msg::AbstractString, key::RSAPrivateKey)

RSA decryption function for strings, expects PKCS#1 v1.5 padding.

ToyPublicKeys.decrypt — Method

decrypt(::pkcs1_v1_5_t, msg::AbstractVector, key::RSAPrivateKey)

RSA decryption function for vectors (arbitrary buffers), expects PKCS#1 v1.5 padding.

ToyPublicKeys.encrypt — Method

encrypt(::pkcs1_v1_5_t,
        msg::Union{AbstractString,AbstractVector},
        key::RSAPublicKey
        ; pad_length=32)

RSA encryption function with PKCS#1 v1.5 padding.

ToyPublicKeys.generate_rsa_key_pair — Method

generate_rsa_key_pair(::pkcs1_v1_5_t, bits::Integer)

RSA key pair constructor (hopefully) according to RFC 2313

ToyPublicKeys.is_probab_prime_p — Method

is_probab_prime_p(n::BigInt, n_tests::Int)

Binding for GMPlib's gmpzprobabprime_p function. Only uses probability based tests as it would equal factoring n` otherwise.

ToyPublicKeys.pad — Method

pad(::pkcs1_v1_5_t, msg::AbstractVector{T}, pad_length=32) where {T<:Base.BitInteger}

Core implementation of the PKCS#1 v1.5 padding.

ToyPublicKeys.pad — Method

pad(::pkcs1_v1_5_t, msg::T, pad_length=32) where {T<:AbstractString}

Wrapper for the core pad function.

ToyPublicKeys.power_crt — Method

power_crt(base::BigInt, pow::BigInt, p::BigInt, q::BigInt)

Wrapper around core implementation, only for generating the parameters if they are not provided.

ToyPublicKeys.power_crt — Method

function power_crt(
    base::BigInt,
    p::BigInt,
    q::BigInt,
    d_p::BigInt,
    d_q::BigInt,
    q_inv::BigInt,
)

CRT for PKCS #1 based parameters.

ToyPublicKeys.power_crt_components — Method

power_crt_components(d::BigInt, p::BigInt, q::BigInt)

Utility function for calculating dth power in p*q mod CRT parameters for PKCS #1.

ToyPublicKeys.power_crt_components — Method

power_crt_components(e::BigInt, d::BigInt, primes::Vector{BigInt})

Utility function for calculating the CRT parameters for PKCS #1.

ToyPublicKeys.rand_prime_for_rsa — Function

rand_prime_for_rsa(bits::Integer, no_gcd_with=big"65537")

Helper function for rsa random primes, since they should satisfy special properties.

ToyPublicKeys.random_bigint_from_range — Method

random_bigint_from_range(bits::Integer)

Custom random big int function since core julia does not yet provide 'proper one' (in my opinion).

ToyPublicKeys.sign — Method

sign(::pkcs1_v1_5_t, msg::AbstractVector, key::RSAPrivateKey; pad_length=32)

Sign AbstractVector (arbitrary buffer using SHA256) with RSA key.

ToyPublicKeys.sign — Method

sign(::pkcs1_v1_5_t, msg::String, key::RSAPrivateKey; pad_length=32)

Sign string with RSA key.

ToyPublicKeys.unpad — Method

unpad(::pkcs1_v1_5_t, msg::AbstractVector{T}) where {T<:Base.BitInteger}

Core implementation for the PKCS#1 v1.5 pad unwrapping.

ToyPublicKeys.unpad — Method

unpad(::pkcs1_v1_5_t, msg::T) where {T<:AbstractString}

Wrapper for the core unpad function.

ToyPublicKeys.verify_signature — Function

verify_signature(::pkcs1_v1_5_t, msg::String, signature::String, key::RSAPublicKey)

Verify the signature.

ToyPublicKeys.RSAKey

ToyPublicKeys.RSAPrivateKey

ToyPublicKeys.RSAPublicKey

ToyPublicKeys.RSADP

ToyPublicKeys.RSAEP

ToyPublicKeys.RSASP1

ToyPublicKeys.RSAStep

ToyPublicKeys.RSAStep

ToyPublicKeys.RSAVP1

ToyPublicKeys.decrypt

ToyPublicKeys.decrypt

ToyPublicKeys.encrypt

ToyPublicKeys.generate_rsa_key_pair

ToyPublicKeys.is_probab_prime_p

ToyPublicKeys.pad

ToyPublicKeys.pad

ToyPublicKeys.power_crt

ToyPublicKeys.power_crt

ToyPublicKeys.power_crt_components

ToyPublicKeys.power_crt_components

ToyPublicKeys.rand_prime_for_rsa

ToyPublicKeys.random_bigint_from_range

ToyPublicKeys.sign

ToyPublicKeys.sign

ToyPublicKeys.unpad

ToyPublicKeys.unpad

ToyPublicKeys.verify_signature